gcp_agents/agent/requirements.txt
flask
google-cloud-aiplatform
streamlit
requests